+++ core/src/lib.rs
@@ -3,15 +3,94 @@
 
 //! Core components and resources to use Heron
 
-use bevy_ecs::Entity;
+use bevy::core::FixedTimestep;
+use bevy::prelude::*;
 
-use bevy_math::Vec3;
+pub use ext::*;
 pub use gravity::Gravity;
 pub use velocity::{AxisAngle, Velocity};
 
+pub mod ext;
 mod gravity;
 pub mod utils;
 mod velocity;
+
+/// Physics stages for user systems. These stages are executed once per physics step.
+///
+/// That usually means they don't run each frame, and may run more than once in a single frame.
+///
+/// In general end-users shouldn't have to deal with these stages directly.
+///
+/// Instead, it is possible to call the [`add_physiscs_system`](ext::AppBuilderExt::add_physics_system) extension function on `AppBuilder`
+/// to register systems that should run during the physics update.
+pub mod stage {
+
+    /// The root **[`Schedule`](bevy::::ecs::Schedule)** stage
+    pub const ROOT: &str = "heron-physics";
+
+    /// A **child** [`SystemStage`](bevy::::ecs::SystemStage) running before each physics step.
+    ///
+    /// Use this stage to modify rigid-body transforms or any other physics component.
+    ///
+    /// **This is not a root stage**. So you cannot simply call `add_system_to_stage` on bevy's app builder.
+    /// Instead consider calling the [`add_physiscs_system`](crate::ext::AppBuilderExt::add_physics_system) extension function.
+    pub const UPDATE: &str = "heron-before-step";
+}
+
+/// Plugin that register stage resources and components.
+///
+/// It does **NOT** enable physics behavior.
+#[derive(Debug, Copy, Clone)]
+pub struct CorePlugin {
+    /// Number of physics step per second. `None` means to run physics step as part of the application update instead.
+    pub steps_per_second: Option<f64>,
+}
+
+impl Default for CorePlugin {
+    fn default() -> Self {
+        Self::from_steps_per_second(60)
+    }
+}
+
+impl CorePlugin {
+    /// Configure how many times per second the physics world needs to be updated
+    ///
+    /// # Panics
+    ///
+    /// Panic if the number of `steps_per_second` is 0
+    #[must_use]
+    pub fn from_steps_per_second(steps_per_second: u8) -> Self {
+        assert!(
+            steps_per_second > 0,
+            "Invalid number of step per second: {}",
+            steps_per_second
+        );
+        Self {
+            steps_per_second: Some(steps_per_second.into()),
+        }
+    }
+}
+
+impl Plugin for CorePlugin {
+    fn build(&self, app: &mut AppBuilder) {
+        app.resources_mut().get_or_insert_with(Gravity::default);
+
+        app.register_type::<Body>()
+            .register_type::<BodyType>()
+            .register_type::<PhysicMaterial>()
+            .register_type::<Velocity>()
+            .add_stage_after(bevy::app::stage::UPDATE, crate::stage::ROOT, {
+                let mut schedule = Schedule::default();
+
+                if let Some(steps_per_second) = self.steps_per_second {
+                    schedule = schedule
+                        .with_run_criteria(FixedTimestep::steps_per_second(steps_per_second))
+                }
+
+                schedule.with_stage(crate::stage::UPDATE, SystemStage::parallel())
+            });
+    }
+}
 
 /// Components that defines a body subject to physics and collision
 ///
@@ -25,7 +104,7 @@ mod velocity;
 ///         .with(Body::Sphere { radius: 1.0 });
 /// }
 /// ```
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Reflect)]
 pub enum Body {
     /// A sphere (or circle in 2d) shape defined by its radius
     Sphere {
@@ -51,6 +130,12 @@ pub enum Body {
     },
 }
 
+impl Default for Body {
+    fn default() -> Self {
+        Self::Sphere { radius: 1.0 }
+    }
+}
+
 /// Component that defines the *type* of rigid body.
 ///
 /// # Example
@@ -64,7 +149,7 @@ pub enum Body {
 ///         .with(BodyType::Static); // Make it static (so that it doesn't move and is not affected by forces like gravity)
 /// }
 /// ```
-#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+#[derive(Debug, Copy, Clone, Eq, PartialEq, Reflect)]
 pub enum BodyType {
     /// A dynamic body is normally affected by physic forces and affect the other bodies normally too.
     ///
@@ -137,40 +222,52 @@ pub enum CollisionEvent {
     Stopped(Entity, Entity),
 }
 
-/// Component that define the [Coefficient of Restitution]
+/// Component that defines the physics properties of the rigid body
 ///
-/// [Coefficient of Restitution]: https://en.wikipedia.org/wiki/Coefficient_of_restitution
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Restitution(f32);
+/// # Example
+///
+/// ```
+/// # use bevy::prelude::*;
+/// # use heron_core::*;
+/// fn spawn(commands: &mut Commands, mut materials: ResMut<Assets<ColorMaterial>>) {
+///     commands.spawn(todo!("Spawn your sprite/mesh, incl. at least a GlobalTransform"))
+///         .with(Body::Sphere { radius: 1.0 }) // Make a body (is dynamic by default)
+///         .with(PhysicMaterial {
+///             restitution: 0.5, // Define the restitution. Higher value means more "bouncy"
+///             density: 2.0, // Define the density. Higher value means heavier.
+///         });
+/// }
+/// ```
+#[derive(Debug, Copy, Clone, PartialEq, Reflect)]
+pub struct PhysicMaterial {
+    /// Coefficient of restitution. Affect how much it "bounces" when colliding other objects.
+    ///
+    /// The higher the value, the more "bouncy".
+    ///
+    /// Typical values are between 0 (perfectly inelastic) and 1 (perfectly elastic)
+    pub restitution: f32,
 
-impl Restitution {
-    /// Perfectly inelastic coefficient, all kinematic energy is lost on collision. (Do not bounce at all)
-    pub const PERFECTLY_INELASTIC: Restitution = Restitution(0.0);
-
-    /// Perfectly elastic coefficient, all kinematic energy is restated in movement. (Very bouncy)
-    pub const PERFECTLY_ELASTIC: Restitution = Restitution(1.0);
-
-    /// Create a new restitution from a coefficient value.
-    #[must_use]
-    pub fn new(coefficient: f32) -> Self {
-        Self(coefficient)
-    }
+    /// Density. Affects how much the body resist to forces.
+    ///
+    /// The higher the value, the heavier.
+    ///
+    /// Value must be greater than 0. Except for sensor and static bodies, in which case the value is ignored.
+    pub density: f32,
 }
 
-impl Default for Restitution {
+impl PhysicMaterial {
+    /// Perfectly inelastic restitution coefficient, all kinematic energy is lost on collision. (Do not bounce at all)
+    pub const PERFECTLY_INELASTIC_RESTITUTION: f32 = 0.0;
+
+    /// Perfectly elastic restitution coefficient, all kinematic energy is restated in movement. (Very bouncy)
+    pub const PERFECTLY_ELASTIC_RESTITUTION: f32 = 1.0;
+}
+
+impl Default for PhysicMaterial {
     fn default() -> Self {
-        Self::PERFECTLY_INELASTIC
-    }
-}
-
-impl From<f32> for Restitution {
-    fn from(value: f32) -> Self {
-        Self(value)
-    }
-}
-
-impl From<Restitution> for f32 {
-    fn from(Restitution(value): Restitution) -> Self {
-        value
+        Self {
+            restitution: Self::PERFECTLY_INELASTIC_RESTITUTION,
+            density: 1.0,
+        }
     }
 }